# [hail] Memory-efficient scan