@tpoterba 2019-06-10
> this should get Laurent cooking with gas.

I rather think it's going to explode, but at least it won't be fizzling out before getting to the meat of densify!