@danking 2019-06-10
blah a bit more complicated than I want it to be, I'll finish this tomorrow.